refactor(commons): move commons to /utils
