+++ src/lib/series/rendering.bars.test.ts
@@ -1,5 +1,5 @@
 import { computeSeriesDomains } from '../../state/utils';
-import { identity } from '../utils/commons';
+import { identity } from '../../utils/commons';
 import { getGroupId, getSpecId, SpecId } from '../utils/ids';
 import { ScaleType } from '../utils/scales/scales';
 import { renderBars } from './rendering';
