+++ src/lib/series/specs.ts
@@ -7,7 +7,7 @@ import {
   BarSeriesStyle,
 } from '../themes/theme';
 import { Accessor } from '../utils/accessor';
-import { Omit, RecursivePartial } from '../utils/commons';
+import { Omit, RecursivePartial } from '../../utils/commons';
 import { AnnotationId, AxisId, GroupId, SpecId } from '../utils/ids';
 import { ScaleContinuousType, ScaleType } from '../utils/scales/scales';
 import { CurveType } from './curves';
