+++ src/lib/utils/scales/scale_band.ts
@@ -1,5 +1,5 @@
 import { scaleBand, scaleQuantize, ScaleQuantize } from 'd3-scale';
-import { clamp } from '../commons';
+import { clamp } from '../../../utils/commons';
 import { ScaleType, Scale } from './scales';
 
 export class ScaleBand implements Scale {
